apps/AI/requirements.txt
# ── Core Data Science ──────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# ── Machine Learning ───────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# ── Visualization ──────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Model Explainability ───────────────────────────────────────────────────
shap>=0.42.0

# ── LLM & Environment ──────────────────────────────────────────────────────
langchain-groq
langchain-core
langchain-community
python-dotenv

# ── Web Framework ──────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
requests

# ── Report Generation (HTML → PDF pipeline) ────────────────────────────────
jinja2                # HTML templating
weasyprint            # Primary HTML → PDF (no external binaries needed)
pdfkit                # Fallback HTML → PDF (requires wkhtmltopdf binary)
